refac: api 이름, 파일구조 수정
