+++ test-web/src/API/fetcher.js
@@ -1,0 +1,16 @@
+const fetcher = (...args) =>
+  fetch(...args).then((res) => {
+    switch (res.status) {
+      //response status가 200 (OK)인 경우, JSON을 파싱한 후 data 반환
+      case 200:
+        return res.json();
+      //response status가 404 (Not Found)인 경우, 에러 throw
+      case 404:
+        throw new Error('No Data Found');
+      //그 외 response status의 경우, JSON을 파싱한 후 data 반환
+      default:
+        return res.json();
+    }
+  });
+
+  export default fetcher
