+++ test-web/src/components/DataDetailPage/DataView.js
@@ -18,9 +18,9 @@ import ApiTable from './tablesComps/apiTable';
 import { TIME_ZONE } from '../../config';
 import Spinner from 'react-bootstrap/Spinner';
 // import update APIs
-import updateHeatedData from '../../API/update/updateHeatedData';
-import updateProbexptData from '../../API/update/updateProbexptData';
-import updateProcessedData from '../../API/update/updateProcessedData';
+import addHeatedData from '../../API/add/addHeatedData';
+import addProbexptData from '../../API/add/addProbexptData';
+import addSensoryProcessedData from '../../API/add/addSensoryProcessedData';
 import RestrictedModal from './restrictedModal';
 // import card
 import QRInfoCard from './cardComps/QRInfoCard';
@@ -125,7 +125,7 @@ function DataView({ dataProps }) {
 
     // 1. 가열육 관능검사 데이터 수정 API POST
     for (let i = 0; i < len; i++) {
-      updateHeatedData(
+      addHeatedData(
         heatInput[i],
         i,
         meatId,
@@ -143,7 +143,7 @@ function DataView({ dataProps }) {
     }
     // 2. 실험실 데이터 수정 API POST
     for (let i = 0; i < len; i++) {
-      updateProbexptData(
+      addProbexptData(
         labInput[i],
         i,
         meatId,
@@ -163,7 +163,7 @@ function DataView({ dataProps }) {
     // 3. 처리육 관능검사 데이터 수정 API POST
     const pro_len = len === 1 ? len : len - 1;
     for (let i = 0; i < pro_len; i++) {
-      updateProcessedData(
+      addSensoryProcessedData(
         processedInput[i],
         processed_data[i],
         processedMinute[i],
